Created BallotFlowTest class to test the IssueBallotFlow and FillBallotFlow classes.  The test verifies that the Ballot is partially signed by running a MockNetwork and starting the flows. The tests verifies that we have an output of a BallotState, that the command being used is IssueBallot and that the participating voter party has signed.
Created BallotContractTest class to verify the existences of the FillOut and Issue commands.  By running a mocknetwork, I was able to add two commands to the ledger (for each command verification).  One of the added commands was
a dummy and would not be added if the test works properly while the other gets added to the ledger once it passes the verification.